Make mod bot lodge reports automatically
