+++ chat.Dockerfile
@@ -15,6 +15,6 @@ COPY \
 RUN : \
   && apt update \
   && apt install -y ffmpeg \
-  && pip install --no-cache-dir -r /app/chat.requirements.txt
+  && pip install --no-cache-dir -r /app/requirements.txt
 
 CMD /app/chat.main.sh
