--- chat.requirements.txt
@@ -1,12 +1,0 @@
-boto3==1.35.99
-confusable-homoglyphs
-erlastic
-fastapi
-lxml
-nltk
-psycopg[binary]
-redis
-regex
-uvicorn
-websockets
-xmltodict
