+++ cron.main.sh
@@ -18,7 +18,7 @@ if [ "${DUO_USE_VENV:-true}" = true ] && [ ! -d venv/cron/ ]
 then
   python3 -m venv venv/cron/
   export PATH=$(readlink -e venv/cron/bin):$PATH
-  python3 -m pip install -r cron.requirements.txt
+  python3 -m pip install -r requirements.txt
 fi
 
 python3 service/cron/__init__.py
