+++ docker-compose.test.yml
@@ -45,7 +45,7 @@ services:
       DUO_SMTP_USER: unused-in-dev-env
       DUO_SMTP_PASS: unused-in-dev-env
 
-      DUO_REPORT_EMAIL: duolicious@example.com 20 a@example.com 10
+      DUO_REPORT_EMAIL: duolicious@example.com 20 a@example.com 0
 
       DUO_DB_HOST: postgres
       DUO_DB_PORT: 5432
@@ -112,6 +112,10 @@ services:
       DUO_SMTP_PORT: 1025
       DUO_SMTP_USER: unused-in-dev-env
       DUO_SMTP_PASS: unused-in-dev-env
+
+      DUO_REPORT_EMAIL: duolicious@example.com 20 a@example.com 0
+
+      DUO_CRON_PROFILE_REPORTER_POLL_SECONDS: 1
 
       DUO_DB_HOST: postgres
       DUO_DB_PORT: 5432
