+++ requirements.txt
@@ -1,16 +1,25 @@
-Flask
-Flask-Limiter
-Pillow
-PyYAML
 blurhash
 boto3==1.35.99
 confusable-homoglyphs
 erlastic
+fastapi
+Flask
 flask-cors
+Flask-Limiter
 gunicorn
+lxml
+nltk
 numpy
+numpy==1.26.4
+onnxruntime
 openai
+Pillow
+pillow-heif
 psycopg[binary]
 pydantic[email]
+PyYAML
 redis
-pillow-heif
+regex
+uvicorn
+websockets
+xmltodict
