+++ service/cron/__init__.py
@@ -6,6 +6,7 @@ from service.cron.nsfwphotorunner import predict_nsfw_photos_forever
 from service.cron.photocleaner import clean_photos_forever
 from service.cron.audiocleaner import clean_audio_forever
 from service.cron.verificationjobrunner import verify_forever
+from service.cron.profilereporter import report_profiles_forever
 import asyncio
 from http.server import SimpleHTTPRequestHandler
 from socketserver import TCPServer
@@ -50,6 +51,8 @@ async def main():
         send_notifications_forever(),
 
         verify_forever(),
+
+        report_profiles_forever(),
 
         check_connections_forever(),
