+++ service/cron/photocleaner/__init__.py
@@ -19,7 +19,7 @@ PHOTO_CLEANER_POLL_SECONDS = int(os.environ.get(
     str(60), # 1 minute
 ))
 
-print('Hello from cron module: photocleaner')
+print(f'Hello from cron module: {__name__}')
 
 async def clean_photos_once():
     params = dict(polling_interval_seconds=PHOTO_CLEANER_POLL_SECONDS)
